Set up basic Om ClojureScript client side application with figwheel
Adds ClojureScript and Om with figwheel as dependencies to the leiningen
project. A figwheel script is provided to start automatic builds of the
client side app for development, and initialize a REPL.

A basic Om.Next app is set up in scavenger.client.main, which currently
renders a list of scavenger items from a local app state.