Set up Datomic free version with very basic schema and sample data
This adds a dependency on the free version of Datomic, with manual
dependency of joda-time to avoid leiningen warning about version ranges
conflicts.

A Datomic transactor properties file is included to easily start up a
Datomic transactor.

In addition it adds a very basic schema EDN for a database, and initial
data EDN for this schema. A Clojure init script is provided to load the
schema and sample data into the Datomic database after starting it up.