Add Google Maps React component from npm and add example usage
This adds a Google Maps React component from npm, adding a foreign
library dependency in figwheel for it to use in ClojureScript.

An example of how to create a factory function for the React component,
similar to Om component factory functions, is provided and a map is
rendered with a default size and center.