Fix basic CORS support
Since our client side app and backend API is currently run under
different ports, XHR from the client app to the backend are cross origin
requests. Thus we need to support the preflight OPTIONS request from the
client and respond with appropriate CORS headers as well as allow origin
on the actual GET request.

Further abstraction of this should definitely be done, as well as
complete the support for POST requests.