Persist new items to the backend before updating local state
When adding a new item we now POST it to the backend API first, getting
a response back with the complete new item including ID, before adding
this to the local state cache.

The client side application is now completely in sync with the backend
data store!

Further improvements should be made though, like using core.async to
avoid callbacks on async requests, and implement optimistic update of
local cache to avoid waiting for API response.