Load ring HTTP handler through figwheel as well
This loads the ring HTTP handler with our Compojure routes for the API
under figwheel as well. The benefit of this is two fold.

First of all we only need to start one server which will both handle our
backend API requests and serving up the client side index.html file.

Secondly this also means that both backend and frontend are behind the
same hostname (localhost:3449), voiding the need for any CORS headers
when doing asynchronous requests from the client side to the API.